refactor(parsers): 三份字幕解析器 isolate 派发脚手架合并为 CueParseDispatch
SRT/ASS/VTT 各自复制 parseStringAsync + _parseStringIsolate +
shouldParseInIsolate（~45 行 x3 逐字同构）。抽 CueParseDispatch.run
统一「大文件切 isolate / 小文件同步」决策；公开 API
（shouldParseInIsolate / largeContentComputeThreshold /
utf8ContentByteLength）保持原签名转发，调用方零改动。
